+++ module/test-dialog.js
@@ -437,14 +437,14 @@ export function oneTestTarget(token, applySize, attackTraits, defenseTraits, tes
           defenseTraits: actor.defenseTraits,
           // then non-vehicle changes
           skills: actor.itemTypes.customSkill.reduce((acc, skill) => {
-            acc[toCamelCase(skill.name)] = { value: skill.system.value, baseAttribute: skill.system.baseAttribute };
+            acc[toCamelCase(skill.name)] = { value: skill.system.value, defenseMod: skill.system.defenseMod, baseAttribute: skill.system.baseAttribute };
             return acc;
           },
             Object.entries(actor.system.skills).reduce((acc, [skillName, skill]) => {
-              acc[skillName] = { value: skill.value, baseAttribute: skill.baseAttribute }
+              acc[skillName] = { value: skill.value, defenseMod: skill.defenseMod ?? 0, baseAttribute: skill.baseAttribute }
               return acc;
             }, {})),
-          attributes: Object.entries(actor.system.attributes).reduce((acc, [key, attr]) => (acc[key] = attr.value, acc), {}),
+          attributes: Object.entries(actor.system.attributes).reduce((acc, [key, attr]) => (acc[key] = attr.value + attr.defenseMod, acc), {}),
           vulnerableModifier: actor.system.statusModifiers.vulnerable,
           darknessModifier: actor.system.statusModifiers.darkness,
           isConcentrating: actor.isConcentrating,